requirements.txt
annotated-types==0.7.0 ; python_version >= "3.10" and python_version < "4.0"
anyio==4.9.0 ; python_version >= "3.10" and python_version < "4.0"
azure-core==1.33.0 ; python_version >= "3.10" and python_version < "4.0"
azure-identity==1.21.0 ; python_version >= "3.10" and python_version < "4.0"
cachetools==5.5.2 ; python_version >= "3.10" and python_version < "4.0"
certifi==2025.1.31 ; python_version >= "3.10" and python_version < "4.0"
cffi==1.17.1 ; python_version >= "3.10" and python_version < "4.0" and platform_python_implementation != "PyPy"
charset-normalizer==3.4.1 ; python_version >= "3.10" and python_version < "4.0"
colorama==0.4.6 ; python_version >= "3.10" and python_version < "4.0" and platform_system == "Windows"
cryptography==44.0.2 ; python_version >= "3.10" and python_version < "4.0"
distro==1.9.0 ; python_version >= "3.10" and python_version < "4.0"
exceptiongroup==1.2.2 ; python_version == "3.10"
google-auth-oauthlib==1.2.1 ; python_version >= "3.10" and python_version < "4.0"
google-auth==2.38.0 ; python_version >= "3.10" and python_version < "4.0"
gspread==6.2.0 ; python_version >= "3.10" and python_version < "4.0"
h11==0.14.0 ; python_version >= "3.10" and python_version < "4.0"
httpcore==1.0.7 ; python_version >= "3.10" and python_version < "4.0"
httplib2==0.22.0 ; python_version >= "3.10" and python_version < "4.0"
httpx==0.28.1 ; python_version >= "3.10" and python_version < "4.0"
idna==3.10 ; python_version >= "3.10" and python_version < "4.0"
jiter==0.9.0 ; python_version >= "3.10" and python_version < "4.0"
joblib==1.4.2 ; python_version >= "3.10" and python_version < "4.0"
msal-extensions==1.3.1 ; python_version >= "3.10" and python_version < "4.0"
msal==1.32.0 ; python_version >= "3.10" and python_version < "4.0"
numpy==2.2.4 ; python_version >= "3.10" and python_version < "4.0"
oauth2client==4.1.3 ; python_version >= "3.10" and python_version < "4.0"
oauthlib==3.2.2 ; python_version >= "3.10" and python_version < "4.0"
openai==1.70.0 ; python_version >= "3.10" and python_version < "4.0"
pyasn1-modules==0.4.2 ; python_version >= "3.10" and python_version < "4.0"
pyasn1==0.6.1 ; python_version >= "3.10" and python_version < "4.0"
pycparser==2.22 ; python_version >= "3.10" and python_version < "4.0" and platform_python_implementation != "PyPy"
pydantic-core==2.33.1 ; python_version >= "3.10" and python_version < "4.0"
pydantic==2.11.2 ; python_version >= "3.10" and python_version < "4.0"
pyjwt[crypto]==2.10.1 ; python_version >= "3.10" and python_version < "4.0"
pyparsing==3.2.3 ; python_version >= "3.10" and python_version < "4.0"
python-dotenv==1.1.0 ; python_version >= "3.10" and python_version < "4.0"
rapidfuzz==3.13.0 ; python_version >= "3.10" and python_version < "4.0"
requests-oauthlib==2.0.0 ; python_version >= "3.10" and python_version < "4.0"
requests==2.32.3 ; python_version >= "3.10" and python_version < "4.0"
rsa==4.9 ; python_version >= "3.10" and python_version < "4"
scikit-learn==1.6.1 ; python_version >= "3.10" and python_version < "4.0"
scipy==1.15.2 ; python_version >= "3.10" and python_version < "4.0"
six==1.17.0 ; python_version >= "3.10" and python_version < "4.0"
sniffio==1.3.1 ; python_version >= "3.10" and python_version < "4.0"
threadpoolctl==3.6.0 ; python_version >= "3.10" and python_version < "4.0"
tqdm==4.67.1 ; python_version >= "3.10" and python_version < "4.0"
typing-extensions==4.13.1 ; python_version >= "3.10" and python_version < "4.0"
typing-inspection==0.4.0 ; python_version >= "3.10" and python_version < "4.0"
urllib3==2.3.0 ; python_version >= "3.10" and python_version < "4.0"